Don't try to commit if there are no changes

+++ doc/deploy-travis.bash
@@ -64,8 +64,12 @@ git clone --quiet --branch=master "git@github.com:flycheck/flycheck.github.io.gi
 cd doc/_deploy
 rake "manual:update[../..,${MANUAL_VERSION}]" 'docs:update[../..]'
 git add --force --all .
-git commit -m "Update manual from flycheck/flycheck@$(git rev-parse --short "${TRAVIS_COMMIT}")"
-git push --force --quiet origin master
+if git diff --quiet; then
+  git commit -m "Update manual from flycheck/flycheck@$(git rev-parse --short "${TRAVIS_COMMIT}")"
+  git push --force --quiet origin master
+else
+  echo "No changes"
+fi
 cd ../..
 
 # shellcheck disable=SC2046
